backend/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-JSON==0.4.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
Werkzeug==3.0.1
